// frontend/src/api/generated/models/getTradeExposureParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fusion Trade Blotter API
 * OpenAPI spec version: 1.0.0
 */
import type { TradeSideInput } from './tradeSideInput';
import type { TradeStatusInput } from './tradeStatusInput';

export type GetTradeExposureParams = {
/**
 * @maxLength 200
 */
search?: string;
status?: TradeStatusInput;
side?: TradeSideInput;
/**
 * @maxLength 120
 */
trader?: string;
/**
 * @maxLength 40
 */
book?: string;
};