// frontend/src/api/generated/models/listTradesParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fusion Trade Blotter API
 * OpenAPI spec version: 1.0.0
 */
import type { SortDirectionInput } from './sortDirectionInput';
import type { TradeSideInput } from './tradeSideInput';
import type { TradeSortKeyInput } from './tradeSortKeyInput';
import type { TradeStatusInput } from './tradeStatusInput';

export type ListTradesParams = {
/**
 * @maxLength 200
 */
search?: string;
status?: TradeStatusInput;
side?: TradeSideInput;
/**
 * @maxLength 120
 */
trader?: string;
/**
 * @maxLength 40
 */
book?: string;
sortBy?: TradeSortKeyInput;
sortDirection?: SortDirectionInput;
/**
 * @maximum 9007199254740991
 * @exclusiveMinimum 0
 */
page?: number;
/**
 * @minimum 1
 * @maximum 200
 */
pageSize?: number;
};